apps/gradio/data_apps_burstbank/requirements.txt
jupyter
python-dotenv

openai
tabulate

pystarburst

gradio
plotly

pyarrow